Refactor util files
