+++ src/visualizer/algorithm/stepbysteptemplate/mergeSortStepByStep.js
@@ -1,5 +1,5 @@
-import { arrayCopy } from '../../sortingvisualizer/util/GeneralUtil';
 import { findIndexToUseInMergeSort } from '../../sortingvisualizer/util/MergeSortUtil';
+import { arrayCopy } from '../../sortingvisualizer/util/ArrayUtil';
 
 /**
  * Generates the step by step template for Merge Sort.
