+++ src/visualizer/sortingvisualizer/component/animationscreen/AnimationScreen.js
@@ -2,18 +2,13 @@ import React, { useContext, useEffect } from 'react';
 import { useTransition } from 'react-spring';
 import './styles.css';
 import { VisualizerStateContext } from '../../Visualizer';
-import {
-  arrayCopy,
-  isBucketSort,
-  isCountingSort,
-  isMergeSort,
-  isRadixSort,
-} from '../../util/GeneralUtil';
+import { isBucketSort, isCountingSort, isMergeSort, isRadixSort } from '../../util/GeneralUtil';
 import CountingSortScreen from './CountingSortScreen';
 import RadixSortScreen from './RadixSortScreen';
 import BucketSortScreen from './BucketSortScreen';
 import MergeSortScreen from './MergeSortScreen';
 import GenericSortScreen from './GenericSortScreen';
+import { arrayCopy } from '../../util/ArrayUtil';
 
 /**
  * The screen which shows the animation for the sorting visualizer.
