+++ src/visualizer/sortingvisualizer/util/GeneralUtil.js
@@ -1,165 +1,8 @@
-import SortingAlgorithms from '../../algorithm/sortingalgorithms/allSorts';
-import SortingAlgorithmsStepByStep from '../../algorithm/stepbysteptemplate/allSortsStepByStep';
-import { generateRadixSortArray } from './RadixSortUtil';
-import { generateBucketSortArray } from './BucketSortUtil';
-import { generateCountSortArray } from './CountingSortUtil';
-import { generateMergeSortArray } from './MergeSortUtil';
-
-/**
- * Resets the array given.
- *
- * @method
- * @category ArrayUtil
- * @param {string} visualizerAlgorithm Algorithm used.
- * @param {Object[]} arr Array to be reset.
- * @returns {Object[]} Array that has been reset.
- */
-export const resetArray = (visualizerAlgorithm, arr) => {
-  if (isRadixOrBucket(visualizerAlgorithm)) {
-    return arrayCopy(arr);
-  }
-
-  return arrayCopy(arr).map((x) => {
-    let tempArrElement = x;
-    if (isCountingSort(visualizerAlgorithm)) {
-      tempArrElement.isShown = true;
-    } else if (isMergeSort(visualizerAlgorithm)) {
-      tempArrElement.isShift = false;
-    } else {
-      tempArrElement.isSwap = false;
-    }
-    return tempArrElement;
-  });
-};
-
-/**
- * Creates a deep copy of the array.
- *
- * @method
- * @category ArrayUtil
- * @param {Object[]} arr Array to be copied.
- * @returns {Object[]} Deep copy of the input array.
- */
-export const arrayCopy = (arr) => {
-  return JSON.parse(JSON.stringify(arr));
-};
-
-/**
- * Retrieves the animation array based on the given array and algorithm selected.
- *
- * @method
- * @category ArrayUtil
- * @param {string} visualizerAlgorithm The current algorithm of the visualizer.
- * @param {Object[]} arrayData Given array.
- * @returns {any[]} An array that contains all the animation steps.
- */
-export const getAnimationArr = (visualizerAlgorithm, arrayData) => {
-  const sortAlgo = SortingAlgorithms[visualizerAlgorithm];
-  return sortAlgo(arrayCopy(arrayData));
-};
-
-/**
- * Gets the step by step text to display to the user.
- *
- * @method
- * @category StepTracingUtil
- * @param {string} visualizerAlgorithm The current algorithm of the visualizer.
- * @param {Object[]} animationArr Animation array.
- * @param {number} idx Index of animation.
- * @param {Object[]} referenceArray Reference array of blocks displayed.
- * @param {Object[]} stackArr Stack array, present only in bucket and radix sort.
- * @returns {string} Step by step text generated.
- */
-export const getStepByStepText = (
-  visualizerAlgorithm,
-  animationArr,
-  idx,
-  referenceArray,
-  stackArr,
-) => {
-  const sortAlgoStepByStep = SortingAlgorithmsStepByStep[visualizerAlgorithm];
-  if (isBucketSort(visualizerAlgorithm)) {
-    return sortAlgoStepByStep(animationArr, idx, stackArr);
-  }
-  return sortAlgoStepByStep(animationArr, idx, referenceArray);
-};
-
-/**
- * Generates a random number within [min, max].
- *
- * @method
- * @param {number} min Min number selected.
- * @param {number} max Max number selected.
- * @returns {number} Random number generated.
- */
-export const generateValue = (min, max) => {
-  return Math.floor(Math.random() * (max - min + 1) + min);
-};
-
-/**
- * Rounds the number to 2 decimal places, if necessary.
- *
- * @method
- * @see {@link https://stackoverflow.com/questions/11832914/round-to-at-most-2-decimal-places-only-if-necessary|Stack Overflow}
- * @param {number} num Input number.
- * @returns {number} Rounded number with maximum 2 decimal places.
- */
-export const roundToTwoDp = (num) => {
-  return +(Math.round(num + 'e+2') + 'e-2');
-};
-
-/**
- * Object used to map the names of each algorithm to its generated array.
- */
-const arrayGenerator = {
-  'Insertion Sort': (size) => generateDefaultArray(size),
-  'Bubble Sort': (size) => generateDefaultArray(size),
-  'Quick Sort': (size) => generateDefaultArray(size),
-  'Shell Sort': (size) => generateDefaultArray(size),
-  'Heap Sort': (size) => generateDefaultArray(size),
-  'Selection Sort': (size) => generateDefaultArray(size),
-  'Merge Sort': (size) => generateMergeSortArray(size),
-  'Counting Sort': (size) => generateCountSortArray(size),
-  'Radix Sort': (size) => generateRadixSortArray(size),
-  'Bucket Sort': (size) => generateBucketSortArray(size),
-};
-
-/**
- * Generates a random array based on the size chosen and the algorithm selected.
- *
- * @method
- * @category ArrayUtil
- * @param {number} size Size of array selected by user.
- * @param {string} visualizerAlgorithm The current algorithm of the visualizer.
- * @returns {Object[]} Random array generated.
- */
-export const generateArray = (size, visualizerAlgorithm) => {
-  return arrayGenerator[visualizerAlgorithm](size);
-};
-
-/**
- * Generates the default array. Used by all sorting algorithms involving swaps only.
- *
- * @category ArrayUtil
- * @param {number} size Size of array selected by user.
- * @returns {Object[]} Random array generated.
- */
-const generateDefaultArray = (size) => {
-  let array = [];
-  for (let i = 0; i < size; i++) {
-    array.push({
-      id: i,
-      height: generateValue(5, 20),
-      isSwap: false,
-    });
-  }
-  return array;
-};
-
 /**
  * Returns true if the algorithm input is counting sort.
  *
  * @method
+ * @category GeneralUtil
  * @param {string} visualizerAlgorithm Input algorithm.
  * @returns {boolean} True if the algorithm is counting sort.
  */
@@ -169,6 +12,7 @@ export const isCountingSort = (visualizerAlgorithm) => visualizerAlgorithm === '
  * Returns true if the algorithm input is radix sort.
  *
  * @method
+ * @category GeneralUtil
  * @param {string} visualizerAlgorithm Input algorithm.
  * @returns {boolean} True if the algorithm is radix sort.
  */
@@ -178,6 +22,7 @@ export const isRadixSort = (visualizerAlgorithm) => visualizerAlgorithm === 'Rad
  * Returns true if the algorithm input is bucket sort.
  *
  * @method
+ * @category GeneralUtil
  * @param {string} visualizerAlgorithm Input algorithm.
  * @returns {boolean} True if the algorithm is bucket sort.
  */
@@ -187,6 +32,7 @@ export const isBucketSort = (visualizerAlgorithm) => visualizerAlgorithm === 'Bu
  * Returns true if the algorithm input is radix or bucket sort.
  *
  * @method
+ * @category GeneralUtil
  * @param {string} visualizerAlgorithm Input algorithm.
  * @returns {boolean} True if the algorithm is radix or bucket sort.
  */
@@ -197,6 +43,7 @@ export const isRadixOrBucket = (visualizerAlgorithm) =>
  * Returns true if the algorithm input is merge sort.
  *
  * @method
+ * @category GeneralUtil
  * @param {string} visualizerAlgorithm Input algorithm.
  * @returns {boolean} True if the algorithm is merge sort.
  */
@@ -206,6 +53,7 @@ export const isMergeSort = (visualizerAlgorithm) => visualizerAlgorithm === 'Mer
  * Returns true if the algorithm input is selection sort.
  *
  * @method
+ * @category GeneralUtil
  * @param {string} visualizerAlgorithm Input algorithm.
  * @returns {boolean} True if the algorithm is selection sort.
  */
@@ -215,6 +63,7 @@ export const isSelectionSort = (visualizerAlgorithm) => visualizerAlgorithm === 
  * Returns true if the algorithm input is quick sort.
  *
  * @method
+ * @category GeneralUtil
  * @param {string} visualizerAlgorithm Input algorithm.
  * @returns {boolean} True if the algorithm is quick sort.
  */
@@ -224,6 +73,7 @@ export const isQuickSort = (visualizerAlgorithm) => visualizerAlgorithm === 'Qui
  * Returns true if the algorithm input is any sort aside from bucket, radix and counting sort.
  *
  * @method
+ * @category GeneralUtil
  * @param {string} visualizerAlgorithm Input algorithm.
  * @returns {boolean} True if the algorithm is any sort aside from bucket, radix and counting sort.
  */
@@ -235,18 +85,3 @@ export const hasLegend = (visualizerAlgorithm) =>
   visualizerAlgorithm === 'Heap Sort' ||
   visualizerAlgorithm === 'Merge Sort' ||
   visualizerAlgorithm === 'Shell Sort';
-
-/**
- * Generates a translation in the x direction to shift the visualizer when the data size is more than 12.
- *
- * @method
- * @param {number} dataSize Data size.
- * @returns {number} Translation in the x direction when the data size is more than 12.
- */
-export const translateXOfVisualizer = (dataSize) => {
-  if (dataSize > 12) {
-    let singleBlockWidth = 200 / dataSize;
-    return (dataSize - 12) * singleBlockWidth;
-  }
-  return 0;
-};
