+++ src/visualizer/sortingvisualizer/util/MergeSortUtil.js
@@ -1,9 +1,11 @@
-import { arrayCopy, generateValue } from './GeneralUtil';
+import { generateValue } from './MathUtil';
+import { arrayCopy } from './ArrayUtil';
 
 /**
  * This is to find the "correct" index to be swapped since the actual array is not being swapped.
  *
  * @method
+ * @category MergeSortUtil
  * @param {Object[]} newTempArr The new input array.
  * @param {number} iIdx I-th Index to be swapped.
  * @param {number} jIdx J-th Index to be swapped.
@@ -24,6 +26,7 @@ export const findIndexToUseInMergeSort = (newTempArr, iIdx, jIdx) => {
  * A sorting execution for Merge Sort.
  *
  * @method
+ * @category MergeSortUtil
  * @param {any[]} currentAnimation The current animation to be executed.
  * @param {Object[]} referenceArray The array holding data for the "blocks".
  * @param {Object[]} historyArr An array holding the data of each states of the reference array.
@@ -46,6 +49,7 @@ export const executeMergeSortForward = (
  * A sorting execution for Merge Sort in the reverse direction.
  *
  * @method
+ * @category MergeSortUtil
  * @param {Object[]} historyArr An array holding the data of each states of the reference array.
  * @param {function} setReferenceArray A set state method for reference array.
  */
@@ -57,6 +61,7 @@ export const executeMergeSortBackward = (historyArr, setReferenceArray) => {
  * A "hackish" method to simulate the swapping in merge swap without actually sorting the array.
  *
  * @method
+ * @category MergeSortUtil
  * @param {Object[]} referenceArray The array holding data for the "blocks".
  * @param {any[]} currentAnimation The current animation to be executed.
  * @returns {Object[]} The resultant array upon the swap animation\.
@@ -101,6 +106,7 @@ export const handleMergeSort = (referenceArray, currentAnimation) => {
  * Generates the random data for merge sort.
  *
  * @method
+ * @category MergeSortUtil
  * @param {number} size The number of blocks to be shown.
  * @returns {Object[]} A array containing the states of each block.
  */
