+++ src/visualizer/sortingvisualizer/util/SwappingAlgoUtil.js
@@ -1,12 +1,12 @@
 // For swapping algorithms
-import { resetArray } from './GeneralUtil';
 import { swap } from '../../algorithm/sortingalgorithms/swap';
+import { resetArray } from './ArrayUtil';
 
 /**
  * Sets the state of the "block" or "oval" to be highlighted.
  *
  * @method
- * @category ArrayUtil
+ * @category SwappingUtil
  * @param {number} firstIdx The first block" or "oval" to be highlighted.
  * @param {number} secondIdx The second block" or "oval" to be highlighted.
  * @param {Object[]} arr The array which the index at {@code firstIdx} and {@code secondIdx} is highlighted
@@ -20,7 +20,7 @@ export const highlight = (firstIdx, secondIdx, arr) => {
  * Swaps the position of the two object in the array.
  *
  * @method
- * @category ArrayUtil
+ * @category SwappingUtil
  * @param {number} firstIdx The first index to be swapped in the array.
  * @param {number} secondIdx The second index to be swapped in the array.
  * @param {Object[]} arr The array which the index at {@code firstIdx} and {@code secondIdx} is swapped and highlighted or just highlighted.
@@ -42,6 +42,7 @@ export const executeSwap = (firstIdx, secondIdx, arr, isSwapOccurring, visualize
  * A generic sorting execution for Insertion, Bubble, Selection, Shell Sort
  *
  * @method
+ * @category SwappingUtil
  * @param {any[]} currentAnimation The current animation to be executed.
  * @param {Object[]} referenceArray The array holding data for the "blocks"
  * @param {string} visualizerAlgorithm The current algorithm of the visualizer
@@ -52,7 +53,7 @@ export const executeGenericSort = (
   currentAnimation,
   referenceArray,
   visualizerAlgorithm,
-  setReferenceArray,
+  setReferenceArray
 ) => {
   let firstIdx = currentAnimation[0];
   let secondIdx = currentAnimation[1];
@@ -62,7 +63,7 @@ export const executeGenericSort = (
     secondIdx,
     referenceArray,
     isSwapOccurring,
-    visualizerAlgorithm,
+    visualizerAlgorithm
   );
   setReferenceArray(arrToUse);
   return arrToUse;
